# [pipeline] avoid local shuffle for scan followed by aggregate