--- a/be/src/exec/vectorized/aggregate/aggregate_blocking_node.cpp
+++ b/be/src/exec/vectorized/aggregate/aggregate_blocking_node.cpp
@@ -172,7 +172,7 @@ std::vector<std::shared_ptr<pipeline::OperatorFactory> > AggregateBlockingNode::
             // The san operator followed by an aggregate operator indicates the aggregating key is the same as
             // the bucket key of this table, so we needn't local shuffle for this case.
             auto* source_op = operators_with_sink[0].get();
-            if (operators_with_sink.size() > 1 || typeid(*source_op) != typeid(pipeline::ScanOperatorFactory)) {
+            if (typeid(*source_op) != typeid(pipeline::ScanOperatorFactory)) {
                 std::vector<ExprContext*> group_by_expr_ctxs;
                 Expr::create_expr_trees(_pool, _tnode.agg_node.grouping_exprs, &group_by_expr_ctxs);
                 operators_with_sink =