### Introduction
The san operator followed by an aggregate operator indicates the aggregating key is the same as the bucket key of this table, so we needn't local shuffle for this case.

After this optimization, in 3 BEs and DOP=8, `sum_sql` costs from 0.65s to 0.30s, and `distinct_sql` costs from 0.61s to 0.27s.
```sql
-- ssb_100g

-- sum_sql
SELECT SUM(lo_ordtotalprice), lo_orderkey FROM lineorder group by lo_orderkey limit 1;

-- distinct_sql
SELECT COUNT (DISTINCT lo_partkey), lo_orderkey FROM lineorder group by lo_orderkey limit 100;
```